Initial FreeBSD support (WIP)

+++ src/video/vulkan/renderer_vulkan.cpp
@@ -2354,7 +2354,7 @@ void teDrawFullscreenTriangle( teShader& shader, teTexture2D& texture, const Sha
     Draw( shader, 0, 0, 0, 0, 3, 0, blendMode, teCullMode::Off, teDepthMode::NoneWriteOff, teTopology::Triangles, teFillMode::Solid, texture.index, teTextureSampler::NearestRepeat, 0, 0, 0, 0 );
 }
 
-void teMapUiMemory( size_t vertexBytes, size_t indexBytes, void** outVertexMemory, void** outIndexMemory )
+void teMapUiMemory( unsigned vertexBytes, unsigned indexBytes, void** outVertexMemory, void** outIndexMemory )
 {
     if (vertexBytes > UiBufferBytes)
     {
